requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
requests>=2.28.0
beautifulsoup4>=4.12.0
jinja2>=3.1.0
statsmodels>=0.14.0
pytz>=2023.3
plyer>=2.1.0
pyzmq>=25.0.0
yfinance>=0.2.0
